feat: make sorting feature more generic.
