+++ src/pages/Lending/types/LendingTypes.ts
@@ -1,14 +1,14 @@
-import { Accounts } from '@/types/lending'
-
 export type OrderByType = 'asc' | 'desc' | 'none'
 
-export type SortIconProps = {
-  id: keyof Accounts
-  sortBy: LendingTableSort
+export type AmbigiousObject = { [key: string]: any }
+
+export type SortIconProps<T extends AmbigiousObject> = {
+  id: keyof T
+  sortBy: LendingTableSort<T>
 }
 
-export type LendingTableSort = {
-  columnName: keyof Accounts | undefined
+export type LendingTableSort<T extends AmbigiousObject> = {
+  columnName: keyof T | undefined
   orderBy: OrderByType
 }
 
@@ -17,12 +17,12 @@ type UnSortableTableHeadType = {
   children: React.ReactNode
 }
 
-type SortableTableHeadType = {
+type SortableTableHeadType<T extends AmbigiousObject> = {
   isSortable: true
-  id: keyof Accounts
+  id: keyof T
   children: React.ReactNode
-  sortBy: LendingTableSort
-  setSortBy: React.Dispatch<React.SetStateAction<LendingTableSort>>
+  sortBy: LendingTableSort<T>
+  setSortBy: React.Dispatch<React.SetStateAction<LendingTableSort<T>>>
 }
 
-export type SortableColumnType = UnSortableTableHeadType | SortableTableHeadType
+export type SortableColumnType<T extends AmbigiousObject> = UnSortableTableHeadType | SortableTableHeadType<T>
